chore: update implementation of sending webhook
